requirements.txt
jax[cuda]
optax
torch
Pillow
tqdm
numpy<2.0.0